formt the code
